enhancement/7 Move feedback request under Dashboard
